requirements.txt
streamlit
langchain-groq
langchain-huggingface
langchain-chroma
chromadb
python-dotenv
sentence-transformers